+++ ft_itoa.c
@@ -6,13 +6,13 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/01/10 13:42:13 by pholster       #+#    #+#                */
-/*   Updated: 2019/03/26 11:41:44 by pholster      ########   odam.nl         */
+/*   Updated: 2019/07/20 19:37:10 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "includes/libft.h"
 
-char	*ft_itoa(long long value)
+char	*ft_itoa(intmax_t value)
 {
 	return (ft_itoa_cbase(value, 10, TRUE));
 }
