+++ threadpool/ft_tasksetinfo.c
@@ -6,15 +6,15 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/04/25 14:54:57 by pholster       #+#    #+#                */
-/*   Updated: 2019/04/25 14:57:14 by pholster      ########   odam.nl         */
+/*   Updated: 2019/07/20 17:18:25 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/threadpool.h"
 
-t_task	*ft_tasksetinfo(t_task *task, void (*f)(), int count, va_list params)
+t_task	*ft_tasksetinfo(t_task *task, void (*f)(), size_t count, va_list params)
 {
-	int		i;
+	size_t		i;
 
 	i = 0;
 	task->fnc = f;
